small improvements to remoting subsystem
